decorator pattern works, only object pool needs to be added

+++ Assets/Scripts/Decorator Pattern/IGun.cs
@@ -5,7 +5,7 @@ public interface IGun
 {
     public GameObject BulletPrefab { get; set; }
     public int Damage { get; set; }
-    public int BulletAmount { get; set; }
+    public float BulletAmount { get; set; }
     public int GrowAmount { get; set; }
     public int ExplodeRadius { get; set; }
     public int BounceAmount { get; set; }
